+++ include/shell/commands.h
@@ -39,9 +39,6 @@ void cmd_clear(const char* args);
 void cmd_echo(const char* args);
 void cmd_secho(const char* args);
 void cmd_memstat(const char* args);
-void cmd_tasks(const char* args);
-void cmd_kill(const char* args);
-void cmd_peek(const char* args);
 void cmd_grep(const char* args);
 
 // fs
@@ -53,5 +50,11 @@ void cmd_mkdir(const char* args);
 void cmd_rm(const char* args);
 void cmd_ls(const char* args);
 void cmd_exec(const char* args);
+
+// tasks
+void cmd_tasks(const char* args);
+void cmd_kill(const char* args);
+void cmd_peek(const char* args);
+void cmd_tlist(const char* args);
 
 #endif